Combat: decompile ~110 more non-leaf functions (0x0209-0x020d combat internals: managers, allocators, GX/palette transfers, event queues, grotto, interrupt-guarded ops)

+++ src/Combat/Main/AppendZeroedEntry0209bd40.cpp
@@ -2,7 +2,7 @@
 #include "std_library_functions.h"
 
 // USA: func_0209bd40
-ARM void AppendZeroedEntry0209bd40(char* obj, short value) {
+ARM void AppendZeroedEntry0209bd40(char* obj, int value) {
     int count = *(int*)(obj + 0xc0);
     if (count >= 6) return;
     memset(obj + count * 0x20, 0, 0x20);
